+++ src/nacos/src/Provider/InstanceProvider.php
@@ -127,15 +127,17 @@ class InstanceProvider extends AbstractProvider
      *     'weight' => 1,
      * ]
      */
-    public function beat(string $serviceName, array $beat = [], ?string $groupName = null, ?string $namespaceId = null, ?bool $ephemeral = null): ResponseInterface
+    public function beat(string $serviceName, array $beat = [], ?string $groupName = null, ?string $namespaceId = null, ?bool $ephemeral = null, bool $lightBeatEnabled = false): ResponseInterface
     {
         return $this->request('PUT', '/nacos/v1/ns/instance/beat', [
             RequestOptions::QUERY => $this->filter([
                 'serviceName' => $serviceName,
+                'ip' => $beat['ip'] ?? null,
+                'port' => $beat['port'] ?? null,
                 'groupName' => $groupName,
                 'namespaceId' => $namespaceId,
                 'ephemeral' => $ephemeral,
-                'beat' => Json::encode($beat),
+                'beat' => ! $lightBeatEnabled ? Json::encode($beat) : '',
             ]),
         ]);
     }